Remove refinerycms-*.rb files to comply with zeitwerk
This is a breaking change, it requires all gems that currently depend on
something in Refinery by requiring the gem's name to update the require
to a path inside the gem.

For example, if a gem had:

    require 'refinerycms-pages'

This has to change to:

    require 'refinery/pages'

Not a huge change, but needs to be done everywhere.